Build the container fresh if the action is not invoked from a tag
Experimental: Attempt to detect whether the action is being invoked from a tag or not.
If the action is invoked by tag, use the prebuilt container to avoid wastefully building the same thing over and over again.
Otherwise (e.g. the action is being invoked from head, maybe as part of a PR presubmit on this repo), build the container on demand.