Build the container fresh every time (#3)
We're building a container now, which is nice for local use. In theory, this action should use the pre-built container instead of making it on the fly every time (which wastes around 25 seconds per run).

However, we really need for PRs to this action to build a fresh Docker container and run it, so we can test it.

I attempted to make a smart solution in #2 that would detect whether the action was being referenced by tag or not, and pass a variable to the composite action 'uses' Docker syntax, but it didn't work.