sample3: add samples for references (#47)
This seems to be the cleanest way to add references, is to use the
metadata block to set the bibliography file and a CSL file. If their is
a CSL we want to standardize on, we can build it INTO the
pandoc/build.sh repo.

This requires option --citeproc and filter "+citation" added to the
pandoc command.